arm64: Dummy out the kernel output folder
All our devices take kernel headers over whatever bionic provides.

In order to do so, they need to be added to the include path and
a dependency on header generation must be added to ensure that
they are there. This does not apply to the emulator because it
uses a prebuilt kernel.

This patch adds a dummy dependency to appease the dependency
management gods.

Change-Id: I19c99829f991f8d6e4a14497fdca1fb6fba64d25